Spring Boot integration working
For the barest definitions of "working", there is a client interceptor
and a server handler interceptor. I wrote an integration test which
sets up a webserver with the signature handler and then executes a
call. I don't think it will return the correct authorization in the
real world, those tests are next...