Refactoring and fixing spring integration tests (still broken)
I've done a bit of moving things around and streamlining interfaces
as I was coding up the examples to make things easier from a user
perspective. Current the integration tests are still broken, it
is becoming apparent that my lack of testing in http-signatures
is a problem. This commit has everything as it is expected to
work once http-signatures is fully tested and working correctly.