Refactoring, cleanup, and testing, also adding RSAPSSA support
There's a lot of cleaning in here and I realized I had made a lot
of changes without a commit so this is a bit of another house cleaning
commit. Until things start to slow down in terms of stability of the
interface, it's probably going to continue to be all over the place.