Adding docker support for echo example servers
Added the spotify docker plugin to the examples project and the project
now generates docker images for the spring boot and netty servers. I haven't
really tested things out, however the build passes and docker images are
produced with the appropriate jars in place. Next steps are the client
docker images and a tomcat example.