Moving from docker plugin to jar plugin / Dockerfile
Docker plugin is awesome for development but it doesn't work right if I'm building
in a sandboxed container environment so I've just switched to a more traditional jar
plugin and Dockerfile build in two steps. There's a lot of duplication between the
docker files right now but I'm just trying to get it working.